move CollectionNode model to tree.py
